Fixed naming conventions, added average latency calculation
